Pricing service working with US 2 and 3

+++ src/main/java/com/tnt/assessment/dto/AggregationDto.java
@@ -8,7 +8,7 @@ import java.util.HashMap;
 @Getter
 @Setter
 public class AggregationDto {
-    private HashMap<String, Float> pricing;
+    private HashMap<String, Double> pricing;
     private HashMap<String, String> track;
     private HashMap<String, String[]> shipments;
 }
